electrical_qna/requirements.txt
Django==4.2.14
google-genai==0.8.0
python-decouple==3.8
mysqlclient==2.2.1
Pillow==10.1.0
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3